requirements.txt
Flask==2.0.1
python-pptx==0.6.21
python-dotenv==0.19.0
PyPDF2==2.10.5
requests==2.28.1
werkzeug==2.0.1
langchain==0.0.267
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
bcrypt==3.2.0